+++ src/main/editor/GridLines.java
@@ -4,6 +4,7 @@ import main.components.Component;
 import main.haspid.Camera;
 import main.haspid.Window;
 import main.renderer.DebugDraw;
+import org.joml.Vector2d;
 import org.joml.Vector2f;
 
 import static main.Configuration.*;
@@ -19,11 +20,11 @@ public class GridLines extends Component {
         int verticalLines = (int)((uProjectionDimension.x * zoom) /  gridSize) + 1;
 
         for(int i = 0; i < verticalLines; i++){
-            DebugDraw.addLine2D(gridLinesIndex, new Vector2f(cam.getPosition().x + (i * (gridSize)), cam.getPosition().y), new Vector2f(cam.getPosition().x + (i * gridSize), cam.getPosition().y + (uProjectionDimension.y * zoom)), gridLinesColor, 1);
+           DebugDraw.addLine2D(gridLinesIndex, new Vector2d(cam.getPosition().x + (i * (gridSize)), cam.getPosition().y), new Vector2d(cam.getPosition().x + (i * gridSize), cam.getPosition().y + (uProjectionDimension.y * zoom)), gridLinesColor, 1);
         }
 
         for(int i = 0; i < horizontalLines; i++){
-            DebugDraw.addLine2D(gridLinesIndex, new Vector2f(cam.getPosition().x, cam.getPosition().y + (i * gridSize)), new Vector2f(cam.getPosition().x + (uProjectionDimension.x * zoom), cam.getPosition().y + (i * gridSize)), gridLinesColor, 1);
+            DebugDraw.addLine2D(gridLinesIndex, new Vector2d(cam.getPosition().x, cam.getPosition().y + (i * gridSize)), new Vector2d(cam.getPosition().x + (uProjectionDimension.x * zoom), cam.getPosition().y + (i * gridSize)), gridLinesColor, 1);
         }
     }
 }
