+++ src/main/renderer/Line2D.java
@@ -1,17 +1,18 @@
 package main.renderer;
 
+import org.joml.Vector2d;
 import org.joml.Vector2f;
 import org.joml.Vector3f;
 
 public class Line2D {
     private int zIndex;
-    private Vector2f to;
+    private Vector2d to;
     private int lifeTime;
     private float stroke;
-    private Vector2f from;
+    private Vector2d from;
     private Vector3f color;
 
-    public Line2D(int zIndex, Vector2f from, Vector2f to, Vector3f color, int lifeTime) {
+    public Line2D(int zIndex, Vector2d from, Vector2d to, Vector3f color, int lifeTime) {
         this.from = from;
         this.to = to;
         this.color = color;
@@ -24,11 +25,11 @@ public class Line2D {
         return lifeTime;
     }
 
-    public Vector2f getFrom() {
+    public Vector2d getFrom() {
         return from;
     }
 
-    public Vector2f getTo() {
+    public Vector2d getTo() {
         return to;
     }
